Drop sixteen more no-op casts that restate their operand's own type
Twelve TUs: eight function-return casts on calls that already return the cast
type (mainGetBit, sbGetPropeller, getWorldMapVoiceoverTimer,
SaveGame_findTransientMapBit), one (s16) on obj->anim.rotX, and seven
*(T*)&field puns that reinterpret a field as the type it is already declared
with (BaddieState.turnRate, CollectibleState.bounceHitFlag,
DimExplosionState.rayMode, DimExplosionFlame.spinAngle/spinSpeed).

Every object byte-identical; main.dol OK.

+++ src/dlls/objects/257_TrickyGuard/TrickyGuard.c
@@ -20,7 +20,7 @@ void TrickyGuard_update(GameObject* obj) {
 
     obj->anim.resetHitboxFlags |= INTERACT_FLAG_DISABLED;
     if (placement->armingGameBit != TRICKYGUARD_GAMEBIT_NONE) {
-        if ((u32)mainGetBit(placement->armingGameBit) == 0) {
+        if (mainGetBit(placement->armingGameBit) == 0) {
             return;
         }
     }
